Added browser cache implementation
If caching is enabled, koa-jade now checks to see if the file has been modified since the last date the request sends over. 

If it has been, it will send back a new modified date and use the compiler as expected. If it hasn't, it will send back a 304 response, prompting the browser to use whatever is in cache.